requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.73.0
audioread==3.1.0
attrs==25.4.0
backoff==2.2.1
banks==2.2.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==43.0.3
dataclasses-json==0.6.7
decorator==4.4.2
defusedxml==0.7.1
deepeval==3.7.0
deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docstring-parser==0.17.0
execnet==2.1.2
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.38.0
google-auth-httplib2==0.2.1
google-genai==1.49.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.0
iniconfig==2.3.0
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
langsmith==0.4.42
lazy-loader==0.4
librosa==0.11.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.8
llama-index-cli==0.5.3
llama-index-core==0.14.8
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.9
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.11.1
llama-parse==0.6.54
llvmlite==0.43.0
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
mypy-extensions==1.1.0
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
ollama==0.6.1
openai==1.109.1
opencv-python==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
packaging==25.0
pandas==2.2.2
pdfminer.six==20251107
Pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
portalocker==3.2.0
posthog==5.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyarrow==18.1.0
pycparser==2.23
pydeck==0.9.1
pydantic==2.11.10
pydantic-core==2.33.2
pydantic-settings==2.12.0
pyfiglet==1.0.4
pygments==2.19.2
pypdf==6.3.0
pytest==8.4.2
pytest-asyncio==1.3.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qdrant-client==1.15.1
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
referencing==0.37.0
rich==13.9.4
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.16.3
sentry-sdk==2.44.0
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soxr==1.0.0
soundfile==0.13.1
SQLAlchemy==2.0.44
streamlit==1.51.0
striprtf==0.0.26
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
tokenizers==0.22.1
tornado==6.5.1
torch==2.8.0
tqdm==4.67.1
tvm==0.17.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typer==0.20.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
whisper==1.1.10
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
